requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
motor==3.6.0
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pymongo==4.9.2
python-dotenv==1.0.1
yarl==1.18.0
